// src/models/customer.ts
// file: src/models/customer.ts
export interface Customer {
  id: number;
  bulan_rek: string;
  bulan: string;
  tgl_baca: string;
  nosbg: string;
  idtarip: number;
  nama: string;
  alamat: string;
  gol: string;
  gol1: string;
  uk_meter: string;
  uk_meter1: string;
  jumlahrp: number;
  retribusi: number;
  danamtr: number;
  biayaadm: number;
  meterai: number;
  stan_lalu: number;
  pembaca: string;
  rayon: string;
  rayonket: string;
  kel: number;
  meter_baca: number;
  pakai: number;
  waktu: string;
  waktu_proses: string;
  tafsir: string;
  ket: string;
  long: string;
  lat: string;
  rata2_3bln: number;
  tgl_download: string;
  nometer: string;
  long1: string;
  lat1: string;
  rata2: number;
}